fix linux print start

+++ docker-builder/build_libs.sh
@@ -34,9 +34,6 @@ build_meson() {
 	if [ -n "${MESON_CROSS_FILE:-}" ]; then
 		cross_args=("--cross-file=$MESON_CROSS_FILE")
 	fi
-	if [ -n "${CROSS_PREFIX:-}" ]; then
-		cross_args+=("--cross-prefix=${CROSS_PREFIX}")
-	fi
 	meson setup build --prefix="$BUILDING_PREFIX" "${cross_args[@]}" --libdir="lib" --buildtype=release --default-library=static "$@"
 	ninja -C build
 	ninja -C build install
@@ -62,12 +59,7 @@ build_autotools() {
 		local -x PTHREAD_CFLAGS=" "
 	fi
 
-	local cross_args=()
-	if [ -n "${CROSS_PREFIX:-}" ]; then
-		cross_args+=("--cross-prefix=${CROSS_PREFIX}")
-	fi
-
-	./configure --prefix="$BUILDING_PREFIX" "${cross_args[@]}" --host="$HOST" --enable-static --disable-shared --with-pic --disable-programs --disable-tools --disable-tests --disable-frontend --disable-examples --disable-docs --disable-cli --disable-unit-tests "$@"
+	./configure --prefix="$BUILDING_PREFIX" --host="$HOST" --enable-static --disable-shared --with-pic --disable-programs --disable-tools --disable-tests --disable-frontend --disable-examples --disable-docs --disable-cli --disable-unit-tests "$@"
 	make -j"$(nproc)"
 	make install
 
@@ -85,9 +77,12 @@ build_make() {
 build_library() {
 	local -x BUILDING_DIR="$1"
 	local -x BUILDING_PREFIX="$2"
+	local BUILD_TIMES_REF="$3"
 	local name
 	name=$(basename "$BUILDING_DIR")
-	echo "--- Compilando $name ---"
+	local start_time
+	start_time=$(date +%s)
+	echo "--- Compilando $name ----------------------------------------------------------------------------------------------------"
 
 	case "$name" in
 	zix) build_meson -Dtests=disabled -Dbenchmarks=disabled ;;
@@ -97,14 +92,16 @@ build_library() {
 	libopenmpt) build_autotools --without-mpg123 --disable-openmpt123 ;;
 	chromaprint)
 		build_cmake -DBUILD_TOOLS=OFF -DBUILD_TESTS=OFF -DFFT_LIB=fftw3 -DFFTW3_DIR="$BUILDING_PREFIX" -DFFTW3_INCLUDE_DIR="$BUILDING_PREFIX/include" -DFFTW3_INCLUDE_DIRS="$BUILDING_PREFIX/include" -DFFTW3_LIBRARY="$BUILDING_PREFIX/lib/libfftw3.a" -DFFTW3_LIBRARIES="$BUILDING_PREFIX/lib/libfftw3.a"
-		cp "$BUILDING_DIR/libchromaprint.pc" "$BUILDING_PREFIX/lib/pkgconfig/"
 		;;
 	sdl2) build_cmake -DSDL_TEST_LIBRARY=OFF -DSDL_TESTS=OFF -DSDL_EXAMPLES=OFF ;;
-	iconv) build_autotools ;;
+	iconv)
+		rm -f "$BUILDING_DIR/configure.ac" "$BUILDING_DIR/autogen.sh" "$BUILDING_DIR/Makefile"
+		build_autotools
+		;;
 	openssl)
 		pushd "$BUILDING_DIR" >/dev/null
 		if [ "$TARGET_OS" == "windows" ]; then
-			./Configure mingw64 --prefix="$BUILDING_PREFIX" --libdir=lib no-shared no-asm
+			./Configure mingw64 --prefix="$BUILDING_PREFIX" --libdir=lib no-shared no-asm no-apps
 		elif [ "$TARGET_OS" == "android" ]; then
 			local ssl_arch=""
 			case "$TARGET_ARCH" in
@@ -115,24 +112,30 @@ build_library() {
 			esac
 			local -x ANDROID_NDK_ROOT="$ANDROID_NDK_HOME"
 			local -x PATH="$TOOLCHAIN/bin:$PATH"
-			./Configure $ssl_arch -D__ANDROID_API__=$API_LEVEL --prefix="$BUILDING_PREFIX" --libdir=lib no-shared -fPIC
+			./Configure $ssl_arch -D__ANDROID_API__=$API_LEVEL --prefix="$BUILDING_PREFIX" --libdir=lib no-shared no-apps no-docs -fPIC
 		else
-			./config --prefix="$BUILDING_PREFIX" --libdir=lib no-shared -fPIC
+			./config --prefix="$BUILDING_PREFIX" --libdir=lib no-shared no-apps no-docs -fPIC
 		fi
 		make -j"$(nproc)"
 		make install_sw
 		popd >/dev/null
 		;;
-	libdavs2 | libxavs2) build_autotools --disable-asm --extra-cflags="-Wno-incompatible-function-pointer-types" ;;
-	libbluray) build_cmake ;;
+	libdavs2 | libxavs2)
+		local BUILDING_DIR="$BUILDING_DIR/build/linux"
+		build_autotools --disable-asm --extra-cflags="-Wno-incompatible-function-pointer-types"
+		;;
+	libbluray) build_meson -Denable_tools=false -Denable_devtools=false -Denable_examples=false ;;
 	zlib) build_autotools ;;
 	libjxl) build_cmake -DJPEGXL_ENABLE_TOOLS=OFF -DJPEGXL_ENABLE_BENCHMARK=OFF -DJPEGXL_ENABLE_EXAMPLES=OFF -DJPEGXL_ENABLE_JNI=OFF -DJPEGXL_ENABLE_SKIA=OFF -DBUILD_TESTING=OFF -DSJPEG_ANDROID_NDK_PATH="${ANDROID_NDK_HOME}" ;;
-	libharfbuzz) build_meson -Dfreetype=enabled -Dtests=disabled -Ddocs=disabled -Dglib=disabled -Dgobject=disabled -Dicu=disabled ;;
+	libharfbuzz) build_meson -Dfreetype=enabled -Dtests=disabled -Ddocs=disabled -Dutilities=disabled -Dgpu=disabled -Dglib=disabled -Dgobject=disabled -Dicu=disabled ;;
 	liblcevc) build_cmake ;;
 	libvpl) build_cmake ;;
 	libgme) build_cmake ;;
 	vulkan-loader) build_cmake ;;
-	libshaderc) build_meson ;;
+	libfreetype) build_meson -Dharfbuzz=disabled ;;
+	fontconfig) build_meson -Dtests=disabled -Dtools=disabled -Ddoc=disabled ;;
+	libsndfile) build_cmake -DENABLE_EXTERNAL_LIBS=OFF ;;
+	libshaderc) build_cmake -DSHADERC_SKIP_TESTS=ON -DSHADERC_SKIP_EXAMPLES=ON ;;
 	libsrt) build_cmake -DENABLE_SHARED=OFF -DENABLE_STATIC=ON -DENABLE_APPS=OFF -DENABLE_TESTING=OFF -DENABLE_UNITTESTS=OFF -DUSE_STATIC_LIBSTDCXX=ON ;;
 	libkvazaar) build_autotools ;;
 	librist)
@@ -144,9 +147,19 @@ build_library() {
 		;;
 	libx265)
 		BUILDING_DIR="$BUILDING_DIR/source"
-		build_cmake -DENABLE_SHARED=OFF -DENABLE_CLI=OFF -DCMAKE_POSITION_INDEPENDENT_CODE=ON
+		local extra_args=()
+		if [ "${TARGET_OS:-}" == "android" ] && [ "$ABI" == "x86" ]; then
+			extra_args+=("-DENABLE_ASSEMBLY=OFF")
+		fi
+		build_cmake -DENABLE_SHARED=OFF -DENABLE_CLI=OFF -DCMAKE_POSITION_INDEPENDENT_CODE=ON "${extra_args[@]}"
 		;;
-	libx264) build_autotools ;;
+	libx264)
+		local extra_args=("--enable-pic")
+		if [ "${TARGET_OS:-}" == "android" ] && [ "$ABI" == "x86" ]; then
+			extra_args+=("--disable-asm")
+		fi
+		build_autotools "${extra_args[@]}"
+		;;
 	lame) build_autotools --disable-decoder ;;
 	libtwolame) build_autotools ;;
 	libuavs3d)
@@ -154,12 +167,24 @@ build_library() {
 		if [ "${TARGET_OS:-}" == "android" ]; then
 			extra_args=("-DCMAKE_THREAD_LIBS_INIT=-lc" "-DCMAKE_HAVE_THREADS_LIBRARY=1" "-DCMAKE_USE_WIN32_THREADS_INIT=0" "-DCMAKE_USE_PTHREADS_INIT=1")
 		fi
-		build_cmake "${extra_args[@]}"
+		build_cmake -DCOMPILE_10BIT=0 "${extra_args[@]}"
 		;;
 	libvorbis) build_cmake -DOGG_LIBRARY="$BUILDING_PREFIX/lib/libogg.a" -DOGG_INCLUDE_DIR="$BUILDING_PREFIX/include" ;;
-	libxml2) build_cmake -DIconv_LIBRARY="$BUILDING_PREFIX/lib/libiconv.a" -DIconv_INCLUDE_DIR="$BUILDING_PREFIX/include" ;;
+	libxml2)
+		if [ "${TARGET_OS:-}" != "linux" ]; then
+			build_cmake -DIconv_LIBRARY="$BUILDING_PREFIX/lib/libiconv.a" -DIconv_INCLUDE_DIR="$BUILDING_PREFIX/include"
+		else
+			build_cmake
+		fi
+		;;
 	expat) build_cmake -DEXPAT_SHARED_LIBS=OFF -DEXPAT_BUILD_EXAMPLES=OFF -DEXPAT_BUILD_TESTS=OFF -DEXPAT_BUILD_DOCS=OFF ;;
-	libpulse) build_meson -Ddatabase=simple -Dtests=false -Dman=false -Dx11=disabled -Ddoxygen=false -Dc_link_args="-L$BUILDING_PREFIX/lib -liconv" ;;
+	libpulse)
+		if [ "${TARGET_OS:-}" != "linux" ]; then
+			build_meson -Ddatabase=simple -Dtests=false -Dman=false -Dx11=disabled -Ddoxygen=false -Dc_link_args="-L$BUILDING_PREFIX/lib -liconv"
+		else
+			build_meson -Ddatabase=simple -Dtests=false -Dman=false -Dx11=disabled -Ddoxygen=false
+		fi
+		;;
 	libpng) build_cmake -DZLIB_LIBRARY="$BUILDING_PREFIX/lib/libz.a" -DZLIB_INCLUDE_DIR="$BUILDING_PREFIX/include" -DPNG_SHARED=OFF -DPNG_TESTS=OFF -DPNG_EXECUTABLES=OFF ;;
 	libvmaf)
 		BUILDING_DIR="$BUILDING_DIR/libvmaf"
@@ -177,7 +202,11 @@ build_library() {
 			x86_64) vpx_targets+=("--target=x86_64-android-gcc") ;;
 			esac
 		fi
-		build_autotools "${vpx_targets[@]}"
+		cd "$BUILDING_DIR"
+		./configure --prefix="$BUILDING_PREFIX" "${vpx_targets[@]}" --enable-pic --disable-examples --disable-unit-tests --disable-tools --disable-docs --disable-shared --enable-static
+		make -j"$(nproc)"
+		make install
+		cd - >/dev/null
 		;;
 	openal) build_cmake -DALSOFT_EXAMPLES=OFF -DALSOFT_UTILS=OFF -DLIBTYPE=STATIC -DCMAKE_EXE_LINKER_FLAGS="-lm" ;;
 	libva) build_meson ;;
@@ -206,7 +235,7 @@ build_library() {
 			cargo_opts+=(--target="$RUST_TARGET")
 			local -x CARGO_TARGET_"$(echo "$RUST_TARGET" | tr '[:lower:]' '[:upper:]' | tr '-' '_')"_LINKER="${CC}"
 		fi
-		cargo cinstall --release --prefix="$BUILDING_PREFIX" --libdir="lib" --library-type=staticlib "${cargo_opts[@]}"
+		cargo cinstall --release --lib --prefix="$BUILDING_PREFIX" --libdir="lib" --library-type=staticlib "${cargo_opts[@]}"
 		if [ -f "$BUILDING_PREFIX/lib/pkgconfig/rav1e.pc" ]; then
 			local content
 			content=$(cat "$BUILDING_PREFIX/lib/pkgconfig/rav1e.pc")
@@ -218,31 +247,45 @@ build_library() {
 		;;
 	librubberband) build_meson ;;
 	libopenh264)
-		local extra_args=()
-		if [ "${TARGET_OS}" == "windows" ]; then
-			extra_args+=("OS=mingw_nt" "ARCH=x86_64")
-		elif [ "${TARGET_OS:-}" == "android" ]; then
-			local openh264_arch=""
-			case "$TARGET_ARCH" in
-			arm64-v8a | arm-v8a) openh264_arch="arm64" ;;
-			armeabi-v7a) openh264_arch="arm" ;;
-			x86) openh264_arch="x86" ;;
-			x86_64) openh264_arch="x86_64" ;;
-			esac
-			extra_args+=("OS=android" "NDKROOT=$ANDROID_NDK_HOME" "TARGET=android-24" "ARCH=$openh264_arch")
+		local args=("PREFIX=$BUILDING_PREFIX" "INCLUDE_PREFIX=$BUILDING_PREFIX/include")
+		if [ "${TARGET_OS:-}" == "android" ]; then
+			args+=("OS=android" "NDKROOT=$ANDROID_NDK_HOME" "TARGET=android-$API_LEVEL" "NDKLEVEL=$API_LEVEL")
+			if [ "$TARGET_ARCH" == "arm64-v8a" ]; then
+				args+=("ARCH=arm64")
+			elif [ "$TARGET_ARCH" == "armeabi-v7a" ]; then
+				args+=("ARCH=arm")
+			elif [ "$TARGET_ARCH" == "x86" ]; then
+				args+=("ARCH=x86" "USE_ASM=No")
+			elif [ "$TARGET_ARCH" == "x86_64" ]; then
+				args+=("ARCH=x86_64")
+			fi
+		elif [ "${TARGET_OS:-}" == "windows" ]; then
+			args+=("OS=mingw_nt" "ARCH=$TARGET_ARCH" "CC=${CC}" "CXX=${CXX}" "AR=${AR}")
 		fi
-		build_make install-static "${extra_args[@]}"
+		build_make libraries "${args[@]}"
+		build_make install "${args[@]}"
 		;;
-	libtheora) build_autotools --disable-asm ;;
+	libtheora)
+		rm -f "$BUILDING_DIR/autogen.sh"
+		cd "$BUILDING_DIR"
+		./configure --prefix="$BUILDING_PREFIX" --host="$HOST" --enable-static --disable-shared --with-pic --disable-examples --disable-spec --disable-asm --disable-maintainer-mode
+		make -j"$(nproc)"
+		make install
+		cd - >/dev/null
+		;;
 	libplacebo) build_meson -Ddemos=false ;;
 	libsoxr) build_cmake -DBUILD_SHARED_LIBS=OFF -DBUILD_TESTS=OFF -DBUILD_EXAMPLES=OFF -DWITH_OPENMP=OFF ;;
 	libssh) build_cmake -DBUILD_SHARED_LIBS=OFF -DWITH_EXAMPLES=OFF -DWITH_SERVER=OFF -DWITH_GSSAPI=OFF -DZLIB_LIBRARY="$BUILDING_PREFIX/lib/libz.a" -DZLIB_INCLUDE_DIR="$BUILDING_PREFIX/include" -DCMAKE_C_FLAGS="-I$BUILDING_PREFIX/include" ;;
 	opencl-icd-loader)
 		build_cmake -DBUILD_SHARED_LIBS=OFF -DOPENCL_ICD_LOADER_HEADERS_DIR="$BUILDING_PREFIX/include"
-		mv "$BUILDING_PREFIX/lib/OpenCL.a" "$BUILDING_PREFIX/lib/libOpenCL.a"
+		if [ -f "$BUILDING_PREFIX/lib/OpenCL.a" ]; then
+			mv "$BUILDING_PREFIX/lib/OpenCL.a" "$BUILDING_PREFIX/lib/libOpenCL.a"
+		fi
 		;;
 	libzvbi)
 		local -x LIBS="-lm"
+		local -x ac_cv_func_malloc_0_nonnull=yes
+		local -x ac_cv_func_realloc_0_nonnull=yes
 		build_autotools --without-x --disable-proxy
 		;;
 	libvvenc)
@@ -250,16 +293,51 @@ build_library() {
 		if [ "${TARGET_OS}" == "windows" ]; then
 			extra_args+=("-DCMAKE_INTERPROCEDURAL_OPTIMIZATION=OFF" "-DVVENC_ENABLE_LTO=OFF")
 		fi
-		build_cmake "${extra_args[@]}" -DBUILD_SHARED_LIBS=OFF
+		if [ "${TARGET_OS:-}" == "android" ] && [ "$ABI" == "x86" ]; then
+			extra_args+=("-DVVENC_ENABLE_X86_SIMD=OFF")
+		fi
+		build_cmake "${extra_args[@]}" -DBUILD_SHARED_LIBS=OFF -DVVENC_ENABLE_WERROR=OFF
 		;;
 	avisynth) build_cmake ;;
+	libass)
+		local extra_args=()
+		if [ "${TARGET_OS:-}" == "android" ] && [[ "$TARGET_ARCH" == *"x86"* ]]; then
+			extra_args=("-Dasm=disabled")
+		fi
+		build_meson "${extra_args[@]}"
+		;;
 	libxvid)
-		build_autotools
-		mv "$BUILDING_PREFIX/lib/xvidcore.a" "$BUILDING_PREFIX/lib/libxvidcore.a"
+		BUILDING_DIR="$BUILDING_DIR/build/generic"
+		local extra_args=()
+		if [ "${TARGET_OS:-}" == "android" ] && [ "$ABI" == "x86" ]; then
+			extra_args+=("--disable-assembly")
+		fi
+		build_autotools "${extra_args[@]}"
+		mv "$BUILDING_PREFIX/lib/xvidcore.a" "$BUILDING_PREFIX/lib/libxvidcore.a" || true
 		;;
 	libzmq) build_cmake -DCMAKE_SYSTEM_VERSION=6.1 -DPOLLER=epoll -DWITH_TLS=OFF -DBUILD_TESTS=OFF -DWITH_DOCS=OFF -DENABLE_DRAFTS=OFF -DBUILD_SHARED=OFF ;;
-	*) echo "Error libreria $name desconocida" && exit 1 ;;
+	libzimg) build_autotools ;;
+	*)
+		if [ -f "$BUILDING_DIR/meson.build" ]; then
+			build_meson
+		elif [ -f "$BUILDING_DIR/CMakeLists.txt" ]; then
+			build_cmake
+		elif [ -f "$BUILDING_DIR/configure" ] || [ -f "$BUILDING_DIR/autogen.sh" ] || [ -f "$BUILDING_DIR/configure.ac" ]; then
+			build_autotools
+		elif [ -f "$BUILDING_DIR/Makefile" ]; then
+			build_make
+		else
+			echo "Error libreria $name desconocida y no se pudo determinar el sistema de compilacion" && exit 1
+		fi
+		;;
 	esac
+	local end_time
+	end_time=$(date +%s)
+	local duration=$((end_time - start_time))
+	if [[ -n "$BUILD_TIMES_REF" ]]; then
+		declare -n ref_times="$BUILD_TIMES_REF"
+		ref_times+=("$duration")
+	fi
 }
 
 compile_linux() {
@@ -269,7 +347,7 @@ compile_linux() {
 	local -x PREFIX="$COMPILATION_DIR/linux_x86_64"
 	local -x BUILD_ROOT="$TEMP_DIR/linux_x86_64"
 	rm -rf "$BUILD_ROOT" && mkdir -p "$BUILD_ROOT" && cp -r "$SRC_ROOT/"* "$BUILD_ROOT"
-	rm -rf "$PREFIX" && mkdir -p "$PREFIX"
+	mkdir -p "$PREFIX"
 	local -x PKG_CONFIG_PATH="$PREFIX/lib/pkgconfig:$PREFIX/lib64/pkgconfig"
 	local -x HOST="x86_64-linux-gnu"
 	local -x CFLAGS="-fPIC -O3 -I$PREFIX/include"
@@ -293,7 +371,6 @@ compile_linux() {
 		"opencl-headers"
 		"opencl-icd-loader"
 		"nv-codec-headers"
-		"iconv"
 		"zlib"
 		"libpng"
 		"libxml2"
@@ -303,7 +380,6 @@ compile_linux() {
 		"fontconfig"
 		"libharfbuzz"
 		"libfribidi"
-		"libshaderc"
 		"libvorbis"
 		"gmp"
 		"lzma"
@@ -365,14 +441,19 @@ compile_linux() {
 		"libpulse"
 		"libdrm"
 	)
+
+	local BUILD_TIMES=()
 	for lib in "${LIBS[@]}"; do
-		if [ -d "$BUILD_ROOT/$lib" ]; then
-			build_library "$BUILD_ROOT/$lib" "$PREFIX"
-		fi
+		build_library "$BUILD_ROOT/$lib" "$PREFIX" "BUILD_TIMES"
 	done
 
 	echo "Librerias compiladas y almacenadas en: $PREFIX"
 	echo "==================== Compilación completada - Linux ====================="
+
+	echo "---- Tiempos de compilación por librería ----"
+	for i in "${!LIBS[@]}"; do
+		printf "%-20s : %d s\n" "${LIBS[$i]}" "${BUILD_TIMES[$i]}"
+	done
 }
 
 compile_windows() {
@@ -382,14 +463,17 @@ compile_windows() {
 	local -x PREFIX="$COMPILATION_DIR/windows_x86_64"
 	local -x BUILD_ROOT="$TEMP_DIR/windows_x86_64"
 	rm -rf "$BUILD_ROOT" && mkdir -p "$BUILD_ROOT" && cp -r "$SRC_ROOT/"* "$BUILD_ROOT"
-	rm -rf "$PREFIX" && mkdir -p "$PREFIX"
+	mkdir -p "$PREFIX"
 
 	local -x PKG_CONFIG_PATH="$PREFIX/lib/pkgconfig:$PREFIX/lib64/pkgconfig"
+	local -x PKG_CONFIG_LIBDIR="$PREFIX/lib/pkgconfig:$PREFIX/lib64/pkgconfig"
 	local -x CROSS_PREFIX="x86_64-w64-mingw32-"
+	"${BUILD_ROOT}/windows-create-pkg-config.sh" "$CROSS_PREFIX" "$PKG_CONFIG_PATH" "$PKG_CONFIG_LIBDIR"
 	local -x CC="${CROSS_PREFIX}gcc"
 	local -x CXX="${CROSS_PREFIX}g++"
 	local -x AR="${CROSS_PREFIX}ar"
 	local -x RANLIB="${CROSS_PREFIX}ranlib"
+	local -x STRIP="${CROSS_PREFIX}strip"
 	local -x RC="${CROSS_PREFIX}windres"
 	local -x HOST="x86_64-w64-mingw32"
 	local -x CFLAGS="-fPIC -O3 -I$PREFIX/include"
@@ -426,7 +510,6 @@ compile_windows() {
 		"fontconfig"
 		"libharfbuzz"
 		"libfribidi"
-		"libshaderc"
 		"libvorbis"
 		"gmp"
 		"lzma"
@@ -483,16 +566,12 @@ compile_windows() {
 		"libzvbi"
 		"libsoxr"
 	)
+
 	for lib in "${LIBS[@]}"; do
-		if [ -d "$BUILD_ROOT/$lib" ]; then
-			build_library "$BUILD_ROOT/$lib" "$PREFIX"
-		fi
+		build_library "$BUILD_ROOT/$lib" "$PREFIX"
 	done
 
-	if [ -f "$SRC_ROOT/windows-pkg-config.sh" ]; then
-		cp "$SRC_ROOT/windows-pkg-config.sh" "$PREFIX/windows-pkg-config.sh"
-		chmod +x "$PREFIX/windows-pkg-config.sh"
-	fi
+	cp "$SRC_ROOT/windows-pkg-config.sh" "$PREFIX/windows-pkg-config.sh"
 
 	echo "Archivos de dependencias precompiladas copiados a /mingw64/"
 	echo "============ Compilación completada - Windows ====================="
@@ -506,7 +585,7 @@ compile_android() {
 	local -x PREFIX="$COMPILATION_DIR/android_$ABI"
 	local -x BUILD_ROOT="$TEMP_DIR/android_$ABI"
 	rm -rf "$BUILD_ROOT" && mkdir -p "$BUILD_ROOT" && cp -r "$SRC_ROOT/"* "$BUILD_ROOT"
-	rm -rf "$PREFIX" && mkdir -p "$PREFIX"
+	mkdir -p "$PREFIX"
 
 	local -x PKG_CONFIG_PATH="$PREFIX/lib/pkgconfig:$PREFIX/lib64/pkgconfig"
 	local -x PKG_CONFIG_LIBDIR="$PKG_CONFIG_PATH"
@@ -622,10 +701,11 @@ compile_android() {
 	)
 
 	for lib in "${LIBS[@]}"; do
-		if [ -d "$BUILD_ROOT/$lib" ]; then
-			build_library "$BUILD_ROOT/$lib" "$PREFIX"
-		fi
+		build_library "$BUILD_ROOT/$lib" "$PREFIX"
 	done
+
+	# Clean up pc files on Android
+	find "$PREFIX/lib/pkgconfig" -name "*.pc" -type f -exec sed -i -e 's/-lpthread//g' -e 's/-lrt//g' -e 's/-lexecinfo//g' -e 's/libexecinfo\.a//g' -e 's/-lunwind//g' -e 's/libunwind\.a//g' {} + || true
 
 	echo "Librerias compiladas y almacenadas en: $PREFIX"
 	echo "==================== Compilación completada - Android $ABI ====================="
